Time: 8 ms (74.52%) | Memory: 17.4 MB (62.50%) - LeetSync

+++ maximum-width-of-binary-tree/maximum-width-of-binary-tree.cpp
@@ -13,7 +13,7 @@ class Solution {
 public:
     int widthOfBinaryTree(TreeNode* root) {
         
-        queue<pair<TreeNode*,long>> q;
+        queue<pair<TreeNode*,int>> q;
         int ind;
 
         q.push({root,0});
